+++ source/AzAuth.PS/Cmdlets/Helpers.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Management.Automation;
 using System.Management.Automation.Language;
+using System.Text.RegularExpressions;
 
 namespace PipeHow.AzAuth;
 
@@ -22,5 +23,23 @@ public class ExistingAccounts : IArgumentCompleter
         catch (Exception) { /* It's fine if we cannot get accounts here */ }
 
         return results;
+    }
+}
+
+public class ValidateCertificatePathAttribute : ValidateArgumentsAttribute
+{
+    protected override void Validate(object arguments, EngineIntrinsics engineIntrinsics)
+    {
+        var path = arguments as string;
+
+        if (!Regex.Match(path, "\\.(pfx|pem)$").Success)
+        {
+            throw new ArgumentException("Only .pfx and .pem files are supported!");
+        }
+
+        if (!File.Exists(path))
+        {
+            throw new ArgumentException($"File '{path}' does not exist.");
+        }
     }
 }